Octopose big restructure to make it pip installable and runnable as commands
