+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include LICENSE
+include octopose/third_party/NuGet.exe
